start/requirements.txt
spacy>=3.7.2
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
mysql-connector-python>=8.2.0
configparser>=6.0.0